+++ example/site/site_links.dart
@@ -1,3 +1,5 @@
 class SiteLinks {
-  static const String productsIndex = '/products/';
+  static const String product = '/product';
+
+  static const String products = '/products';
 }