+++ example/views/error404.html.g.dart
@@ -5,9 +5,14 @@ import 'layout.html.g.dart';
 class error404_html {
   String render(HttpRequest request) {
     final out = StringBuffer();
-    out.write('Page not found\n');
+    out.write('<h1>\n');
+    out.write('    Error 404\n');
+    out.write('</h1>\n');
+    out.write('<p>\n');
+    out.write('    Page not found\n');
+    out.write('</p>\n');
     final layout = layout_html();
-    layout.addTag(HtmlTag('meta', {'description': 'Error 404'}));
+    layout.addMeta({'name': 'description', 'content': 'Error 404'});
     layout.render(out, request, statusCode: 404);
     return out.toString();
   }