+++ tool/build_readme.dart
@@ -1,6 +1,6 @@
 import 'build.dart';
 
 Future<void> main(List<String> args) async {
-  final files = ['tool/readme'];
+  final files = ['tool/readme_md'];
   await build(files);
 }